\name{dengue_hoang}
\alias{dengue_hoang}
\docType{data}
\title{
Example validation dataset for SIGORA 
}
\description{
Up-regulated genes in dengue fever.
}
\source{
Hoang L.T., Lynn D.J., Henn M., Birren B.W., Lennon N.J., Lee P.T., Duong K.T., Nguyen T.T., Mai L.N., Farrar J.J., Hibberd M.L., Simmons C.P. (2010). An Early Whole-blood Transcriptional Signature Associated with Progression to Dengue Shock Syndrome in Vietnamese Children. Journal of Virology 84(24), 12982-12994.}
\examples{
\dontrun{
sigs(dengue_hoang,'k',1,level=2)
}
}
\keyword{datasets}