\name{ecm_randall_b6}
\alias{ecm_randall_b6}
\docType{data}
\title{
Example validation dataset for SIGORA 
}
\description{
Significantly up-regulated genes expressed in the brain tissue of B6 mice subjected to experimental cerebral malaria.
}
\source{
Randall, L. M., Amante, F. H., McSweeney, K. a, Zhou, Y., Stanley, A. C., Haque, A., et al. (2008). Common strategies to prevent and modulate experimental cerebral malaria in mouse strains with different susceptibilities. Infection and Immunity, 76(7), 3312-20.
}
\examples{
\dontrun{
sigs(ecm_randall_b6,'k',1,level=2)
}
}
\keyword{datasets}