% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_package.R
\docType{data}
\name{idmap}
\alias{idmap}
\title{Identifier mappings for protein coding genes.}
\source{
www.ensembl.org/biomart/martview
}
\description{
A mapping table for ENSEMBL, ENTREZ and gene names(HGNC/MGI symbols) of
Human and mouse protein coding gene.
}
\examples{

data(idmap)
head(idmap)

}
\keyword{datasets}
