% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_package.R
\docType{data}
\name{kegH}
\alias{kegH}
\title{Pathway GPS data, extracted from KEGG repository (Human).}
\source{
<http://www.genome.jp/kegg/pathway.html>
}
\description{
KEGG human pathway GPS data, extracted by \code{makeGPS}, default settings.
This data can be used by \code{sigora} to preform signature
overrepresenation.
}
\examples{

data(kegH)
str(kegH)

}
\references{
Kanehisa, M., Goto, S., Sato, Y., Furumichi, M., & Tanabe, M.
2012. \dQuote{KEGG for integration and interpretation of large-scale
molecular data sets.} \emph{Nucleic Acids Research} \bold{40}(D1).
}
\seealso{
\code{\link{makeGPS}}, \code{\link{sigora}} , \code{\link{reaH}}
}
\keyword{datasets}
