% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_package.R
\docType{data}
\name{kegM}
\alias{kegM}
\title{Pathway GPS data, extracted from KEGG repository (Mouse).}
\source{
<http://www.genome.jp/kegg/pathway.html>
}
\description{
KEGG mouse pathway GPS data, extracted by \code{makeGPS}, default settings.
This data can be used by \code{sigora} to preform signature
overrepresenation.
}
\examples{

data(kegM)
## maybe str(kegM) ; plot(kegM) ...

}
\references{
Kanehisa, M., Goto, S., Sato, Y., Furumichi, M., & Tanabe, M.
2012. \dQuote{KEGG for integration and interpretation of large-scale
molecular data sets.} \emph{Nucleic Acids Research} \bold{40}(D1).
}
\keyword{datasets}
