% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load data when lazyLoad true.}
\usage{
load_data(datastr, package = "sigora")
}
\arguments{
\item{datastr}{name of datasets}

\item{package}{default sigora}
}
\description{
load data when lazyLoad true.
}
