\name{reaH}
\alias{reaH}
\docType{data}
\title{
Pathway GPS data, extracted from the Reactome repository (Human).
}
\description{
Reactome human pathway GPS data, extracted by \code{makeGPS}, default settings.
This data can be used by \code{sigora} to preform signature overrepresenation.
}
\usage{data("reaH")}
\references{
Matthews, L., Gopinath, G., Gillespie, M., Caudy, M., Croft, D.,  et al. 2009.
\dQuote{Reactome knowledgebase of human biological pathways and processes.}
 \emph{Nucleic acids research} \bold{37}(Database issue).
}
\source{
<http://www.reactome.org/>
}
\examples{
data(reaH)
## maybe str(reaH) ;  ...
}
\keyword{datasets}
