\name{smallpox_rubins}
\alias{smallpox_rubins}
\docType{data}
\title{
Example validation dataset for SIGORA 
}
\description{
Human one to one orthologs of upregulated genes in host (macaques) transcriptional response during smallpox (variola virus) infection.
}
\source{
Rubins, K. H., Hensley, L. E., Jahrling, P. B., Whitney, A. R., Geisbert, T. W., Huggins, J. W., Owen, A., et al. (2004). The host response to smallpox: analysis of the gene expression program in peripheral blood cells in a nonhuman primate model. Proceedings of the National Academy of Sciences of the United States of America, 101(42), 15190-5.
}
\examples{
\dontrun{
idb_converter(smallpox_rubins)
}
}
\keyword{datasets}