\name{idmap}
\alias{idmap}
\docType{data}
\title{
Identifier mappings for protein coding genes.
}
\description{
A mapping table for ENSEMBL, ENTREZ and gene names(HGNC/MGI symbols) of
Human and mouse protein coding gene.
}
\usage{data("idmap")}
\source{
www.ensembl.org/biomart/martview
}
\examples{
data(idmap)
head(idmap) 
}
\keyword{datasets}
