% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load and return data when lazyLoad false insted of using data(datastr)}
\usage{
load_data(datastr, package = "sigora")
}
\arguments{
\item{datastr}{name of datasets}

\item{package}{default sigora}
}
\value{
returns the data
}
\description{
load and return data when lazyLoad false insted of using data(datastr)
}
