% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_package.R
\docType{data}
\name{reaH}
\alias{reaH}
\title{Pathway GPS data, extracted from the Reactome repository (Human).}
\source{
<http://www.reactome.org/>
}
\description{
Reactome human pathway GPS data, extracted by \code{makeGPS}, default
settings. This data can be used by \code{sigora} to preform signature
overrepresenation.
}
\examples{

data(reaH)
## maybe str(reaH) ;  ...

}
\references{
Matthews, L., Gopinath, G., Gillespie, M., Caudy, M., Croft, D.,
et al. 2009. \dQuote{Reactome knowledgebase of human biological pathways and
processes.} \emph{Nucleic acids research} \bold{37}(Database issue).
}
\keyword{datasets}
