% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_package.R
\docType{data}
\name{reaM}
\alias{reaM}
\title{Pathway GPS data, extracted from Reactome repository (Mouse).}
\source{
<http://www.reactome.org/>
}
\description{
Reactome mouse pathway GPS data, extracted by \code{makeGPS}, default
settings. This data can be used by \code{sigora} to preform signature
overrepresenation.
}
\examples{

data(reaM)
str(reaM)

}
\references{
Matthews, L., Gopinath, G., Gillespie, M., Caudy, M., Croft, D.,
et al. 2009. \dQuote{Reactome knowledgebase of human biological pathways and
processes.} \emph{Nucleic acids research} \bold{37}(Database issue).
}
\seealso{
\code{\link{makeGPS}}, \code{\link{sigora}} , \code{\link{kegM}}
}
\keyword{datasets}
